Return ICD-10-PCS Code as a gt Table

Return ICD-10-PCS Code as a gt Table

Usage

gt_pcs2(x)

Arguments

x

an object of class pcs.

Value

description

Examples

pcs("0G9000Z") |> gt_pcs2()

  

0G9000Z  Drainage of Pituitary Gland with Drainage Device, Open Approach
    

  

    ICD-10-PCS 2024
  

    

1
	Section
	0
	Medical and Surgical

2
	Body System
	G
	Endocrine System

3
	Operation
	9
	Drainage

4
	Body Part
	0
	Pituitary Gland

5
	Approach
	0
	Open

6
	Device
	0
	Drainage Device

7
	Qualifier
	Z
	No Qualifier

gt_pcs2(pcs("0016070"))

  

0016070  Bypass Cerebral Ventricle to Nasopharynx with Autologous Tissue Substitute, Open Approach
    

  

    ICD-10-PCS 2024
  

    

1
	Section
	0
	Medical and Surgical

2
	Body System
	0
	Central Nervous System and Cranial Nerves

3
	Operation
	1
	Bypass

4
	Body Part
	6
	Cerebral Ventricle

5
	Approach
	0
	Open

6
	Device
	7
	Autologous Tissue Substitute

7
	Qualifier
	0
	Nasopharynx