Look up ICD-10-PCS Codes or Tables

Look up ICD-10-PCS Codes or Tables

Usage

pcs(x = NULL)

Arguments

x

an alphanumeric character vector, can be 3 to 7 characters long.

Value

a tibble

Examples

pcs("0G9")
#> ── `procedural::pcs()` ────────────────────────────────────── ICD-10-PCS 2024 ──
#>   
#>   Selected: 0G9
#> 
#> 1. [0] Section: Medical and Surgical
#> 2. [G] Body System: Endocrine System
#> 3. [9] Operation: Drainage
#> 
#> → Select Body Part
#> 
#> • 0 >=> Pituitary Gland
#> • 1 >=> Pineal Body
#> • 2 >=> Adrenal Gland, Left
#> • 3 >=> Adrenal Gland, Right
#> • 4 >=> Adrenal Glands, Bilateral
#> • 6 >=> Carotid Body, Left
#> • 7 >=> Carotid Body, Right
#> • 8 >=> Carotid Bodies, Bilateral
#> • 9 >=> Para-aortic Body
#> • B >=> Coccygeal Glomus
#> • C >=> Glomus Jugulare
#> • D >=> Aortic Body
#> • F >=> Paraganglion Extremity
#> • G >=> Thyroid Gland Lobe, Left
#> • H >=> Thyroid Gland Lobe, Right
#> • K >=> Thyroid Gland
#> • L >=> Superior Parathyroid Gland, Right
#> • M >=> Superior Parathyroid Gland, Left
#> • N >=> Inferior Parathyroid Gland, Right
#> • P >=> Inferior Parathyroid Gland, Left
#> • Q >=> Parathyroid Glands, Multiple
#> • R >=> Parathyroid Gland

pcs("0G90")
#> ── `procedural::pcs()` ────────────────────────────────────── ICD-10-PCS 2024 ──
#>   
#>   Selected: 0G90
#> 
#> 1. [0] Section: Medical and Surgical
#> 2. [G] Body System: Endocrine System
#> 3. [9] Operation: Drainage
#> 4. [0] Body Part: Pituitary Gland
#> 
#> → Select Approach
#> 
#> • 0 >=> Open
#> • 3 >=> Percutaneous
#> • 4 >=> Percutaneous Endoscopic

pcs("0G900")
#> ── `procedural::pcs()` ────────────────────────────────────── ICD-10-PCS 2024 ──
#>   
#>   Selected: 0G900
#> 
#> 1. [0] Section: Medical and Surgical
#> 2. [G] Body System: Endocrine System
#> 3. [9] Operation: Drainage
#> 4. [0] Body Part: Pituitary Gland
#> 5. [0] Approach: Open
#> 
#> → Select Device
#> 
#> • 0 >=> Drainage Device
#> • Z >=> No Device

pcs("0G9000")
#> ── `procedural::pcs()` ────────────────────────────────────── ICD-10-PCS 2024 ──
#>   
#>   Selected: 0G9000
#> 
#> 1. [0] Section: Medical and Surgical
#> 2. [G] Body System: Endocrine System
#> 3. [9] Operation: Drainage
#> 4. [0] Body Part: Pituitary Gland
#> 5. [0] Approach: Open
#> 6. [0] Device: Drainage Device
#> 
#> → Select Qualifier
#> 
#> • Z >=> No Qualifier

pcs("0G9000Z")
#> $code
#> [1] "0G9000Z"
#> 
#> $description
#> [1] "Drainage of Pituitary Gland with Drainage Device, Open Approach"
#> 
#> $axes
#> # A tibble: 7 × 4
#>    axis name        value label               
#>   <int> <chr>       <chr> <chr>               
#> 1     1 Section     0     Medical and Surgical
#> 2     2 Body System G     Endocrine System    
#> 3     3 Operation   9     Drainage            
#> 4     4 Body Part   0     Pituitary Gland     
#> 5     5 Approach    0     Open                
#> 6     6 Device      0     Drainage Device     
#> 7     7 Qualifier   Z     No Qualifier        
#> 
#> $definitions
#> # A tibble: 2 × 2
#>   label    definition                                                           
#>   <chr>    <chr>                                                                
#> 1 Drainage Taking or letting out fluids and/or gases from a body part. The qual…
#> 2 Open     Cutting through the skin or mucous membrane and any other body layer…
#> 
#> $includes
#> # A tibble: 1 × 5
#>   section axis  name      label           includes                              
#>   <chr>   <chr> <chr>     <chr>           <chr>                                 
#> 1 0       4     Body Part Pituitary Gland Adenohypophysis, Hypophysis, Neurohyp…
#> 
#> attr(,"class")
#> [1] "pcs"  "list"